fix: move form routes and components to src/app structure
Move form implementation to match repository reorganization:
- src/routes/forms → src/app/routes/forms
- src/components/flex-form-* → src/app/components/flex-form-*

Update all import paths in routes, components, server, and tests.
