+++ src/app/components/flex-form-field/index.tsx
@@ -1,5 +1,5 @@
 import type { FC } from 'hono/jsx'
-import type { DataRequirement, FieldEntry } from '../../types/models'
+import type { DataRequirement, FieldEntry } from '../../../types/models'
 import { Checkbox } from '../flex-checkbox'
 import { DatePicker } from '../flex-date-picker'
 import { ErrorMessage } from '../flex-error-message'
