+++ src/app/server.tsx
@@ -4,9 +4,9 @@ import { getBasePath, resolveUrl } from '../lib/base-path'
 import { Layout } from './components/flex-layout'
 import catalog from './routes/catalog/index'
 import { createFormRouter } from './routes/forms/index'
-import { InMemoryFormSessionGateway } from './services/form-session'
-import { InMemorySubmissionGateway } from './services/submission'
-import { testDataSpec, testFormSpec } from '../test/forms/fixtures'
+import { InMemoryFormSessionGateway } from '../services/form-session'
+import { InMemorySubmissionGateway } from '../services/submission'
+import { testDataSpec, testFormSpec } from '../../test/forms/fixtures'
 
 const basePath = getBasePath()
 const app = new Hono().basePath(basePath)
